+++ Singularity/Singularity/Levels/Tutorial.cs
@@ -4,7 +4,7 @@ using Microsoft.Xna.Framework.Content;
 using Microsoft.Xna.Framework.Graphics;
 using Singularity.Manager;
 using Singularity.Map;
-using Singularity.Platform;
+using Singularity.Platforms;
 using Singularity.Screen.ScreenClasses;
 using Singularity.Units;
