+++ Singularity/Singularity/PlatformActions/ProduceResource.cs
@@ -1,10 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
 using Singularity.Map;
+using Singularity.Platforms;
 using Singularity.Resources;
 using Singularity.Units;
 
-namespace Singularity.Platform
+namespace Singularity.PlatformActions
 {
 
     public class ProduceWellResource : APlatformAction
