+++ Singularity/Singularity/Platforms/CommandCenter.cs
@@ -4,10 +4,11 @@ using System.Runtime.Serialization;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using Singularity.Manager;
+using Singularity.PlatformActions;
 using Singularity.Resources;
 using Singularity.Units;
 
-namespace Singularity.Platform
+namespace Singularity.Platforms
 {
     [DataContract()]
     class CommandCenter: PlatformBlank
