+++ Singularity/Singularity/Platforms/Mine.cs
@@ -4,31 +4,33 @@ using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using Singularity.Manager;
 using Singularity.Map;
+using Singularity.PlatformActions;
 using Singularity.Resources;
 
-namespace Singularity.Platform
+namespace Singularity.Platforms
 {
     [DataContract]
-    class Quarry : PlatformBlank
+    class Mine : PlatformBlank
     {
         [DataMember]
         private const int PlatformWidth = 144;
         [DataMember]
-        private const int PlatformHeight = 127;
+        private const int PlatformHeight = 187;
+
         [DataMember]
         private Director mDirector;
 
-        public Quarry(Vector2 position, Texture2D platformSpriteSheet, Texture2D baseSprite, ResourceMap resource, ref Director dir): base(position, platformSpriteSheet, baseSprite, new Vector2(position.X + PlatformWidth / 2f, position.Y + PlatformHeight - 36))
+        public Mine(Vector2 position, Texture2D spritesheet, Texture2D basesprite, ResourceMap resource, ref Director dir) : base(position, spritesheet, basesprite, new Vector2(position.X + PlatformWidth / 2f, position.Y + PlatformHeight - 36))
         {
             mDirector = dir;
             dir.GetDistributionManager.Register(this, false);
-            //Add possible Actions in this array
+
             mIPlatformActions = new IPlatformAction[2];
-            mIPlatformActions[0] = new ProduceQuarryResource(this, resource);
+            mIPlatformActions[0] = new ProduceMineResource(this, resource);
             //Something like "Hello Distributionmanager I exist now(GiveBlueprint)"
             //Add Costs of the platform here if you got them.
             mCost = new Dictionary<EResourceType, int>();
-            mType = EPlatformType.Quarry;
+            mType = EPlatformType.Mine;
             mSpritename = "Dome";
             AbsoluteSize = SetPlatfromDrawParameters();
         }
