+++ Singularity/Singularity/Platforms/Quarry.cs
@@ -4,32 +4,32 @@ using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using Singularity.Manager;
 using Singularity.Map;
+using Singularity.PlatformActions;
 using Singularity.Resources;
 
-namespace Singularity.Platform
+namespace Singularity.Platforms
 {
     [DataContract]
-    class Mine : PlatformBlank
+    class Quarry : PlatformBlank
     {
         [DataMember]
         private const int PlatformWidth = 144;
         [DataMember]
-        private const int PlatformHeight = 187;
-
+        private const int PlatformHeight = 127;
         [DataMember]
         private Director mDirector;
 
-        public Mine(Vector2 position, Texture2D spritesheet, Texture2D basesprite, ResourceMap resource, ref Director dir) : base(position, spritesheet, basesprite, new Vector2(position.X + PlatformWidth / 2f, position.Y + PlatformHeight - 36))
+        public Quarry(Vector2 position, Texture2D platformSpriteSheet, Texture2D baseSprite, ResourceMap resource, ref Director dir): base(position, platformSpriteSheet, baseSprite, new Vector2(position.X + PlatformWidth / 2f, position.Y + PlatformHeight - 36))
         {
             mDirector = dir;
             dir.GetDistributionManager.Register(this, false);
-
+            //Add possible Actions in this array
             mIPlatformActions = new IPlatformAction[2];
-            mIPlatformActions[0] = new ProduceMineResource(this, resource);
+            mIPlatformActions[0] = new ProduceQuarryResource(this, resource);
             //Something like "Hello Distributionmanager I exist now(GiveBlueprint)"
             //Add Costs of the platform here if you got them.
             mCost = new Dictionary<EResourceType, int>();
-            mType = EPlatformType.Mine;
+            mType = EPlatformType.Quarry;
             mSpritename = "Dome";
             AbsoluteSize = SetPlatfromDrawParameters();
         }
