+++ Singularity/Singularity/Singularity.csproj
@@ -104,22 +104,22 @@
     <Compile Include="Input\InputManager.cs" />
     <Compile Include="Input\KeyEvent.cs" />
     <Compile Include="Input\MouseAction.cs" />
-    <Compile Include="Platform\CommandCenter.cs" />
-    <Compile Include="Platform\EnergyFacility.cs" />
-    <Compile Include="Platform\EPlatformType.cs" />
-    <Compile Include="Platform\Factory.cs" />
-    <Compile Include="Platform\Junkyard.cs" />
-    <Compile Include="Platform\Mine.cs" />
-    <Compile Include="Platform\Quarry.cs" />
-    <Compile Include="Platform\Storage.cs" />
-    <Compile Include="Platform\Well.cs" />
-    <Compile Include="Platform\IPlatformAction.cs" />
-    <Compile Include="Platform\Road.cs" />
+    <Compile Include="Platforms\CommandCenter.cs" />
+    <Compile Include="Platforms\EnergyFacility.cs" />
+    <Compile Include="Platforms\EPlatformType.cs" />
+    <Compile Include="Platforms\Factory.cs" />
+    <Compile Include="Platforms\Junkyard.cs" />
+    <Compile Include="Platforms\Mine.cs" />
+    <Compile Include="Platforms\Quarry.cs" />
+    <Compile Include="Platforms\Storage.cs" />
+    <Compile Include="Platforms\Well.cs" />
+    <Compile Include="PlatformActions\IPlatformAction.cs" />
+    <Compile Include="Platforms\Road.cs" />
     <Compile Include="Property\ICollider.cs" />
     <Compile Include="Property\IRevealing.cs" />
     <Compile Include="Resources\EResourceType.cs" />
-    <Compile Include="Platform\Action.cs" />
-    <Compile Include="Platform\PlatformBlank.cs" />
+    <Compile Include="PlatformActions\Action.cs" />
+    <Compile Include="Platforms\PlatformBlank.cs" />
     <Compile Include="Screen\EOptionScreenState.cs" />
     <Compile Include="Screen\ITransitionableMenu.cs" />
     <Compile Include="Screen\IWindowItem.cs" />
@@ -170,8 +170,8 @@
     <Compile Include="Utils\Optional.cs" />
     <Compile Include="Utils\Pair.cs" />
     <Compile Include="Utils\Triple.cs" />
-    <Compile Include="Platform\PlatformActionState.cs" />
-    <Compile Include="Platform\ProduceResource.cs" />
+    <Compile Include="PlatformActions\PlatformActionState.cs" />
+    <Compile Include="PlatformActions\ProduceResource.cs" />
     <Compile Include="Resources\Resource.cs" />
     <Compile Include="Screen\WindowObject.cs" />
   </ItemGroup>
