+++ Singularity/Singularity/Units/GeneralUnit.cs
@@ -5,7 +5,8 @@ using Microsoft.Xna.Framework.Graphics;
 using Singularity.Graph;
 using Singularity.Libraries;
 using Singularity.Manager;
-using Singularity.Platform;
+using Singularity.PlatformActions;
+using Singularity.Platforms;
 using Singularity.Property;
 using Singularity.Resources;
 using Singularity.Utils;
